another round of refactoring
cleaned up messages, comments and code without changing the logic
