+++ src/test/kotlin/com/sourcegraph/cody/KeyboardShortcutTest.kt
@@ -1,0 +1,74 @@
+package com.sourcegraph.cody
+
+import com.intellij.openapi.actionSystem.ActionManager
+import com.intellij.openapi.actionSystem.KeyboardShortcut
+import com.intellij.openapi.util.SystemInfoRt
+import com.intellij.testFramework.fixtures.BasePlatformTestCase
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.junit.runners.JUnit4
+import javax.swing.KeyStroke
+
+@RunWith(JUnit4::class)
+class KeyboardShortcutTest : BasePlatformTestCase() {
+
+  // TODO: Need UI tests for testing Send Code Instructions hotkey.
+  // There is no action id; it is a handled by a key listener in the EditCommandPrompt dialog.
+  companion object {
+    // These shortcuts in this declaration should match the table here:
+    // https://linear.app/sourcegraph/document/keyboard-shortcuts-for-cody-on-jetbrains-2f8747a22530
+    // The key shortcuts are bound in plugin.xml unless otherwise specified below.
+    private val keysToActionIds =
+        mapOf(
+            "cody.acceptAutocompleteAction" to arrayOf("TAB", "TAB"), // Windows, macOS
+
+            "cody.command.Explain" to arrayOf("ctrl alt 1", "ctrl alt 1"),
+            "cody.command.Smell" to arrayOf("ctrl alt 2", "ctrl alt 2"),
+
+            // This command also handles cody.inlineEditRetryAction:
+            "cody.editCodeAction" to arrayOf("ctrl alt COMMA", "ctrl alt ENTER"),
+            "cody.documentCodeAction" to arrayOf("ctrl alt H", "ctrl alt H"),
+            "cody.testCodeAction" to arrayOf("ctrl alt G", "ctrl alt G"),
+
+            // Handles cody.inlineEditUndoAction, inlineEditCancelAction, & inlineEditDismissAction:
+            "cody.editCancelOrUndoAction" to arrayOf("ctrl alt BACK_SPACE", "ctrl alt BACK_SPACE"),
+
+            "cody.editShowDiffAction" to arrayOf("ctrl alt D", "ctrl alt K"),
+            "cody.inlineEditAcceptAction" to arrayOf("ctrl shift EQUALS", "ctrl shift EQUALS"),
+
+            "cody.newChat" to arrayOf("ctrl alt 0", "ctrl alt 0"),
+            "cody.openChat" to arrayOf("ctrl alt 9", "ctrl alt 9"),
+
+            "cody.triggerAutocomplete" to arrayOf("control alt P", "control alt P"),
+
+            "sourcegraph.openFindPopup" to arrayOf("alt S", "alt S"),
+        )
+  }
+
+  @Test
+  fun testShortcuts() {
+    for ((actionId, expectedKeys) in keysToActionIds) {
+      val (windowsKey, macKey) = expectedKeys
+
+      val action = ActionManager.getInstance().getAction(actionId)
+      checkNotNull(action) { "No action found for actionId: $actionId" }
+
+      val actualKeyStrokes =
+          action.shortcutSet.shortcuts
+              .mapNotNull { shortcut ->
+                when (shortcut) {
+                  is KeyboardShortcut -> shortcut.firstKeyStroke
+                  else -> null
+                }
+              }
+              .toTypedArray()
+
+      val key = if (SystemInfoRt.isMac) macKey else windowsKey
+      val expectedKey = KeyStroke.getKeyStroke(key)
+
+      assertTrue(
+          "Incorrect keybinding for $actionId: expected $expectedKey, got ${actualKeyStrokes.contentToString()}",
+          actualKeyStrokes.contains(expectedKey))
+    }
+  }
+}
